+++ CHANGELOG.md
@@ -18,6 +18,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 * default comment symbols for `Config::cargo` changed to `#`
 * Ctrl+C now prints the summary of the tests that were run before Ctrl+C is pressed
+* `//@only-target` and `//@only-host` are now separated from the triple substring by a `:` instead of a `-`
+* `//@only-64bit` is now `//@only-bitwidth: 64`
 
 ### Removed